Forget current when making new tenant current...
I'm not 100% sure that this change is a good idea.  It makes sense to me for what I *think* would be good, in general.  I'm not sure, however, because this is my first jaunt down Multi-Tenancy Lane.  ;-)

Anyway, this small change does two things on calls to `makeCurrent()`:
- First, it will return early if the tenant is already the current.
- Second, it will call `forgetCurrent()` for the "old" tenant.

Please let me know what's good about this and, more importantly, what's bad with this and why.  :-)

Thx.